Add the first functional test checking the browser's title.
Functional tests let us see how the application functions from the user's point of view.

User story: How the user might work with a feature and how the app should respond to them.

Terminology: acceptance == functional == end-to-end TESTS

Functional tests can be a sort of specification for the application: it tends to track a user story. Writing fts, we can write the commands first, and replace it later with code.